feat: Track the person with a target marker #12
